Code Review fixes & first custom hook
